Search PATH again if an executable has disappeared
This addresses a problem that can come up when ./bin is on your PATH.
As an example: you run `rake` when working on a Rails project with
binstubs, and bash caches the path to the executable as `./bin/rake`.
If you then change directory and try to run `rake` again, bash will
reuse the cached path and run `./bin/rake`, which now doesn't exist.

This option tells bash to check before every invocation whether the
executable still exists, and search the full PATH again if it doesn't.